FakeReader now in delta_tracker.py
